fix(web-search): friendly copy for message-less native failures + dedupe web_search test harness (ATL-727)
